requirements.txt
pandas==1.5.3
avro-python3==1.10.2
requests==2.31.0
